requirements.txt
pillow
pyheif
inquirer
tqdm